backend/requirements.txt
django>=4.2.0,<4.3.0
djangorestframework>=3.14.0,<3.15.0
psycopg2-binary>=2.9.6,<2.10.0
python-dotenv>=1.0.0,<1.1.0
drf-yasg>=1.21.5,<1.22.0
django-cors-headers>=4.0.0,<4.1.0
djangorestframework-simplejwt>=5.2.2,<5.3.0
requests>=2.31.0
whitenoise>=6.5.0
channels>=4.0.0
django-environ>=0.11.2
django-filter>=23.3
Pillow>=10.1.0
